+++ common_spec_types.go
@@ -1,5 +1,10 @@
 package druid
 
+import (
+	"encoding/json"
+	"fmt"
+)
+
 // BitmapFactory is a field of IndexSpec.
 type BitmapFactory struct {
 	Type string `json:"type"`
@@ -94,7 +99,7 @@ type SpatialDimension struct {
 type TransformSet []Transform
 
 // DimensionSet is a unique set of druid datasource dimensions(labels).
-type DimensionSet []any
+type DimensionSet []DimensionSpec
 
 // Dimension is a typed definition of a datasource dimension.
 type Dimension struct {
@@ -119,6 +124,19 @@ type DimensionsSpec struct {
 	IncludeAllDimensions bool                   `json:"includeAllDimensions,omitempty"`
 	UseSchemaDiscovery   bool                   `json:"useSchemaDiscovery,omitempty"`
 }
+
+// DimensionSpec is a single dataset dimension that can be represented by a typed Dimension or a string value.
+type DimensionSpec struct {
+	Value any
+}
+
+// QueryGranularitySpec is an umbrella type for different representations of query granularity, can be string or
+// QueryGranularity value.
+type QueryGranularitySpec struct {
+	Value any
+}
+
+// QueryGranularity is a typed representation of query granularity.
 type QueryGranularity struct {
 	Type string `json:"type,omitempty"`
 }
@@ -127,12 +145,11 @@ type QueryGranularity struct {
 // partitioning, truncating timestamps, time chunk segmentation or roll-up.
 // https://druid.apache.org/docs/latest/ingestion/ingestion-spec#granularityspec
 type GranularitySpec struct {
-	Type               string `json:"type"`
-	SegmentGranularity string `json:"segmentGranularity,omitempty"`
-	// TODO: this field is problematic as depending on value druid returns string or object
-	QueryGranularity any      `json:"queryGranularity,omitempty"`
-	Rollup           bool     `json:"rollup"`
-	Intervals        []string `json:"intervals,omitempty"`
+	Type               string                `json:"type"`
+	SegmentGranularity string                `json:"segmentGranularity,omitempty"`
+	QueryGranularity   *QueryGranularitySpec `json:"queryGranularity,omitempty"`
+	Rollup             bool                  `json:"rollup,omitempty"`
+	Intervals          []string              `json:"intervals,omitempty"`
 }
 
 // AutoScalerConfig is part of IOConfig that controls ingestion auto-scaling.
@@ -322,4 +339,42 @@ type InputSource struct {
 // https://druid.apache.org/docs/latest/ingestion/ingestion-spec#transformspec
 type TransformSpec struct {
 	Transforms TransformSet `json:"transforms"`
+}
+
+func (g *QueryGranularitySpec) UnmarshalJSON(b []byte) error {
+	var str string
+	if err := json.Unmarshal(b, &str); err == nil {
+		g.Value = str
+		return nil
+	}
+
+	var qg QueryGranularity
+	if err := json.Unmarshal(b, &qg); err == nil {
+		g.Value = qg
+		return nil
+	}
+	return fmt.Errorf("unsupported query granularity: %s", b)
+}
+
+func (g *QueryGranularitySpec) MarshalJSON() ([]byte, error) {
+	return json.Marshal(&g.Value)
+}
+
+func (g *DimensionSpec) UnmarshalJSON(b []byte) error {
+	var str string
+	if err := json.Unmarshal(b, &str); err == nil {
+		g.Value = str
+		return nil
+	}
+
+	var qg Dimension
+	if err := json.Unmarshal(b, &qg); err == nil {
+		g.Value = qg
+		return nil
+	}
+	return fmt.Errorf("unsupported dimension value: %s", b)
+}
+
+func (g *DimensionSpec) MarshalJSON() ([]byte, error) {
+	return json.Marshal(&g.Value)
 }
